requirements.txt
numpy
pandas
tensorflow-cpu
opencv-python-headless>=4.1.0
mtcnn>=0.1.0
split_folders
flask
werkzeug
ultralytics
imageio-ffmpeg
pillow
scikit-learn